heroku_app/requirements.txt
pandas==1.1.1
numpy==1.19.1
streamlit==0.65.2
nltk==3.5
re==2.2.1
lexical-diversity==0.1.1
newspaper3k==0.2.8
tldextract==2.2.3
syltippy==1.0
scikit-learn==0.23.1
xgboost==1.1.1
spacy==2.3.2
https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-2.3.1/es_core_news_lg-2.3.1.tar.gz#egg=en_core_news_lg==2.3.1